门户集成和权限控制接口

+++ src/main/java/io/dfjinxin/modules/sys/dao/SysMenuDao.java
@@ -14,6 +14,7 @@ import org.apache.ibatis.annotations.Mapper;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Map;
 
 /**
  * 菜单管理
@@ -28,7 +29,7 @@ public interface SysMenuDao extends BaseMapper<SysMenuEntity> {
 	 * 根据父菜单，查询子菜单
 	 * @param parentId 父菜单ID
 	 */
-	List<SysMenuEntity> queryListParentId(Long parentId);
+	List<SysMenuEntity> queryListParentId(int parentId);
 	
 	/**
 	 * 获取不包含按钮的菜单列表
@@ -38,6 +39,8 @@ public interface SysMenuDao extends BaseMapper<SysMenuEntity> {
 	/**
 	 * 查询用户的所有菜单ID
 	 */
-	List<Long> queryAllMenuId(Long userId);
+	List<Integer> queryAllMenuId(int userId);
+
+
 
 }
